Repositoryクラスのジェネリクスのミスの修正

+++ src/main/java/com/example/project/domain/InquiryRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InquiryRepository extends JpaRepository<InquiryEntity, Integer> {
+public interface InquiryRepository extends JpaRepository<InquiryEntity, Long> {
 }
